requirements.txt
autopep8==1.6.0
click==8.0.3
dnspython==2.2.1
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.2
Flask-MySQL==1.5.2
Flask-WTF==1.0.1
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mongoengine==0.24.1
paho-mqtt==1.6.1
pycodestyle==2.8.0
pymongo==4.1.1
python-decouple==3.5
python-dotenv==0.19.1
six==1.16.0
toml==0.10.2
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1

